refactor: Update blog page to display unique tags
This commit refactors the blog page to display unique tags for each blog post. It adds a new section in the sidebar that lists all the unique tags associated with the blog posts. Each tag is displayed as a pill-shaped button that links to the corresponding tag page. Additionally, a "View all" link is added at the bottom of the section to navigate to the page displaying all the blog categories.